% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adjust_age_sex_indirectly.R
\name{calculate_adjust_age_sex_indirectly}
\alias{calculate_adjust_age_sex_indirectly}
\title{Title}
\usage{
calculate_adjust_age_sex_indirectly(table, mode, fraction, conf_level, ...)
}
\arguments{
\item{conf_level}{}
}
\description{
Title
}
