% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_disease_cluster.R
\name{calculate_disease_cluster}
\alias{calculate_disease_cluster}
\title{Title}
\usage{
calculate_disease_cluster(model = "spatial", table, ...)
}
\arguments{
\item{model}{}
}
\description{
Title
}
