% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_disease_map.R
\name{calculate_disease_map}
\alias{calculate_disease_map}
\title{Title}
\usage{
calculate_disease_map(model = "spatial", table, graph_file_path, ...)
}
\arguments{
\item{model}{}
}
\description{
Title
}
