% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_forecasting.R
\name{calculate_forecasting}
\alias{calculate_forecasting}
\title{Title}
\usage{
calculate_forecasting(
  table,
  observation_end_date = "2008-01-01",
  prediction_end_date = "2009-08-01",
  ...
)
}
\arguments{
\item{prediction_end_date}{}
}
\description{
Title
}
