% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohort_analysis_table.R
\name{get_cohort_analysis_table}
\alias{get_cohort_analysis_table}
\title{Title}
\usage{
get_cohort_analysis_table(
  model = "spatial",
  conn_info,
  cdm_database_schema,
  result_database_schema,
  target_cohort_definition_id,
  outcome_cohort_definition_id,
  cohort_start_date,
  cohort_end_date,
  time_at_risk_start_date,
  time_at_risk_end_date,
  time_at_risk_end_date_panel,
  ...
)
}
\arguments{
\item{model}{}
}
\description{
Title
}
