% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohort_list_table.R
\name{get_cohort_list_table}
\alias{get_cohort_list_table}
\title{Title}
\usage{
get_cohort_list_table(conn_info, result_database_schema, ...)
}
\arguments{
\item{result_database_schema}{}
}
\description{
Title
}
